Building volume after startup.
I just realized that the reason nothing appears to exist in the home directory is because the mount replaces it. This service will hopefully allow us to do the setup after the volume has been mounted. By doing it this way, the volume itself will contain the new home contents.